Fix bad helm values output
