+++ requirements.txt
@@ -1,6 +1,6 @@
 simpledi>=0.2
 PyYAML>=5.1
-himl==0.11.0
+himl>=0.11.1
 fastjsonschema==2.14.*
 termcolor>=1.1.0
 kubeconfig>=1.1.1
