+++ src/kompos/helpers/himl_helper.py
@@ -8,8 +8,6 @@
 # OF ANY KIND, either express or implied. See the License for the specific language
 # governing permissions and limitations under the License.
 
-import os
-
 from himl.config_generator import ConfigProcessor
 
 from kompos import display
@@ -115,38 +113,3 @@ class HierarchicalConfigGenerator:
             command += " --multi-line-string"
 
         return command
-
-
-def discover_compositions(config_path):
-    path_params = dict(split_path(x) for x in config_path.split('/'))
-
-    composition_type = path_params.get(COMPOSITION_KEY, None)
-    if not composition_type:
-        raise Exception("No composition detected in path.")
-
-    # Check if single composition selected
-    composition = path_params.get(composition_type, None)
-    if composition:
-        return [composition], {composition: config_path}
-
-    # Discover composition paths
-    paths = dict()
-    compositions = []
-    for subpath in os.listdir(config_path):
-        if composition_type + "=" in subpath:
-            composition = split_path(subpath)[1]
-            paths[composition] = os.path.join(config_path, "{}={}".format(composition_type, composition))
-            compositions.append(composition)
-
-    return compositions, paths
-
-
-def sorted_compositions(compositions, composition_order, reverse=False):
-    result = list(filter(lambda x: x in compositions, composition_order))
-    return tuple(reversed(result)) if reverse else result
-
-
-def split_path(value, separator='='):
-    if separator in value:
-        return value.split(separator)
-    return [value, ""]
